refac(91026): Extrai task gerar_notificacao_inicio_periodo_prestacao_de_contas_async
